// src/types/enum.ts
/** @format */

export enum ERoutes {
  Search = "/search",
  Signup = "/sign-up",
  Dashboard = "/dashboard",
  Profile = "/profile",
  Orders = "/orders",
  Summary = "/summary",
  Faq = "/faq/",
  Visit = "/visit",
  Campaign = "/campaign",
  Upgrade = "/upgrade",
  Login = "/login",
  Success = "/success",
  EditSearchOrder = "/orders/editSearchTask",
  EditVisitOrder = "/orders/editVisitTask",
  Calendar = "/orders/calendar",
  VerifyEmail = "/verify-email",
  ResetPassword = "/forget_password",
}